add initial support for suggesting a root extraction folder in 7zFM
this makes it so that when you press Extract (or Copy To since it's the same thing)
in the FileManager, it will now determine if a `NameOfArchive.ext` should be
extracted in `.` or `./NameOfArchive`. without this, archives without a root/
encompassing folder will extract all the contents in the containing dir by default.
Regardless of OS and use case, this is usually the desired case. There are rare
instances where you want the extracted files to essentially emplace a folder
structure (i.e. installing things to /usr/local/lib, etc)

In addition, this implementation will not suggest `./NameOfArchive` if the user
is extracting a folder called `NameOfArchive` (preventing
`./NameOfArchive/NameOfArchive/...ArchiveContents`). This scenario happens
when a user archives a single folder, which is usually the case to prevent
the former issue from happening in certain archive extractors.